pyproject.toml
[tool.poetry]
name = "nlp-app-samples"
version = "0.1.0"
description = ""
authors = ["李国冬 <liguodongiot@163.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
torch = "^1.9.1"
torchvision = "^0.10.1"
transformers = "^4.11.0"
pandas = "^1.3.3"
dataclasses-json = "^0.5.6"
scikit-learn = "^1.0"
progressbar = "^2.5"
scipy = "^1.8.1"
hydra-core = "^1.2.0"
datasets = "^2.4.0"
jieba = "^0.42.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
